[DX] Move MissingOptionException and AddTranslatableFieldSubscriber to EnrichBundle
