+++ src/Pim/Bundle/TranslationBundle/Form/Type/TranslatableFieldType.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\TranslationBundle\Form\Type;
 
 use Pim\Bundle\CatalogBundle\Helper\LocaleHelper;
-use Pim\Bundle\TranslationBundle\Form\Subscriber\AddTranslatableFieldSubscriber;
+use Pim\Bundle\EnrichBundle\Form\Subscriber\AddTranslatableFieldSubscriber;
 use Pim\Bundle\UserBundle\Context\UserContext;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Exception\InvalidConfigurationException;
