--- src/Pim/Bundle/TranslationBundle/Tests/Entity/InvalidTranslation.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace Pim\Bundle\TranslationBundle\Tests\Entity;
-
-/**
- * Invalid Translation
- *
- * @author    Gildas Quemener <gildas@akeneo.com>
- * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-class InvalidTranslation
-{
-}
