--- src/Pim/Bundle/TranslationBundle/Tests/Entity/Item.php
@@ -1,109 +1,0 @@
-<?php
-
-namespace Pim\Bundle\TranslationBundle\Tests\Entity;
-
-use Akeneo\Component\Localization\Model\AbstractTranslation;
-use Akeneo\Component\Localization\Model\TranslatableInterface;
-
-/**
- * Test class
- *
- * @author    Romain Monceau <romain@akeneo.com>
- * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-class Item implements TranslatableInterface
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function setLocale($locale)
-    {
-        $this->locale = $locale;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getTranslation($locale = null)
-    {
-        $locale = ($locale) ? $locale : $this->locale;
-        foreach ($this->translations as $translation) {
-            if ($translation->getLocale() == $locale) {
-                return $translation;
-            }
-        }
-
-        $translationClass = $this->getTranslationFQCN();
-        $translation      = new $translationClass();
-        $translation->setForeignKey($this);
-
-        return $translation;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getTranslations()
-    {
-        return $this->translations;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function addTranslation(AbstractTranslation $translation)
-    {
-        if (!$this->translations->contains($translation)) {
-            $this->translations->add($translation);
-        }
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function removeTranslation(AbstractTranslation $translation)
-    {
-        $this->translations->removeElement($translation);
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getTranslationFQCN()
-    {
-        return 'Pim\Bundle\TranslationBundle\Tests\Entity\ItemTranslation';
-    }
-
-    /**
-     * Get name
-     *
-     * @return string
-     */
-    public function getName()
-    {
-        $translated = $this->getTranslation()->getName();
-
-        return ($translated !== '') ? $translated : '['.$this->getCode().']';
-    }
-
-    /**
-     * Set name
-     *
-     * @param string $name
-     *
-     * @return string
-     */
-    public function setName($name)
-    {
-        $this->getTranslation()->setName($name);
-
-        return $this;
-    }
-}
