--- src/Pim/Bundle/TranslationBundle/Tests/Entity/ItemTranslation.php
@@ -1,44 +1,0 @@
-<?php
-
-namespace Pim\Bundle\TranslationBundle\Tests\Entity;
-
-use Akeneo\Component\Localization\Model\AbstractTranslation;
-
-/**
- * Item Translation
- *
- * @author    Romain Monceau <romain@akeneo.com>
- * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-class ItemTranslation extends AbstractTranslation
-{
-    /**
-     * @var string
-     */
-    protected $name;
-
-    /**
-     * Set name
-     *
-     * @param string $name
-     *
-     * @return AbstractTranslation
-     */
-    public function setName($name)
-    {
-        $this->name = $name;
-
-        return $this;
-    }
-
-    /**
-     * Get name
-     *
-     * @return string $name
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
-}
