--- src/Pim/Bundle/TranslationBundle/Tests/Unit/Form/Subscriber/AddTranslatableFieldSubscriberTest.php
@@ -1,506 +1,0 @@
-<?php
-
-namespace Pim\Bundle\TranslationBundle\Tests\Unit\Form\Subscriber;
-
-use Pim\Bundle\TranslationBundle\Form\Subscriber\AddTranslatableFieldSubscriber;
-
-/**
- * Test related class
- *
- * @author    Romain Monceau <romain@akeneo.com>
- * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-class AddTranslatableFieldSubscriberTest extends \PHPUnit_Framework_TestCase
-{
-    /**
-     * @var \Symfony\Component\Form\FormFactory
-     */
-    protected $formFactory;
-
-    /**
-     * @var \Symfony\Component\Form\Form
-     */
-    protected $form;
-
-    protected $localeConfig = [
-        'locales' => [
-            'fr_FR' => ['label' => 'fr_FR'],
-            'en_US' => ['label' => 'en_US']
-        ]
-    ];
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function setUp()
-    {
-        $this->formFactory = $this->getFormFactoryMock();
-        $this->form        = $this->getFormMock();
-    }
-
-    /**
-     * Data provider with missing option
-     *
-     * @return array
-     */
-    public static function missingOptionDataProvider()
-    {
-        return [
-            'miss_translation_class' => [
-                []
-            ],
-            'miss_entity_class' => [
-                ['translation_class' => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\ItemTranslation']
-            ],
-            'miss_field' => [
-                [
-                    'translation_class' => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\ItemTranslation',
-                    'entity_class'      => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\Item'
-                ]
-            ],
-            'miss_locales' => [
-                [
-                    'translation_class' => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\ItemTranslation',
-                    'entity_class'      => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\Item',
-                    'field'             => 'name'
-                ]
-            ],
-            'miss_widget' => [
-                [
-                    'translation_class' => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\ItemTranslation',
-                    'entity_class'      => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\Item',
-                    'field'             => 'name',
-                    'locales'           => ['fr_FR', 'en_US']
-                ]
-            ],
-            'miss_required_locale' => [
-                [
-                    'translation_class' => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\ItemTranslation',
-                    'entity_class'      => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\Item',
-                    'field'             => 'name',
-                    'locales'           => ['fr_FR', 'en_US'],
-                    'widget'            => 'text'
-                ]
-            ]
-        ];
-    }
-
-    /**
-     * Test with missing option
-     *
-     * @param array $options
-     *
-     * @dataProvider missingOptionDataProvider
-     * @expectedException \Pim\Bundle\TranslationBundle\Exception\MissingOptionException
-     */
-    public function testMissingOptionException(array $options)
-    {
-        $target             = $this->getTargetedClass($options);
-        $translatableEntity = $this->getTranslatableEntityMock();
-        $event              = $this->getEventMock($this->form, $translatableEntity, []);
-
-        $target->preSetData($event);
-        $target->submit($event);
-        $target->postBind($event);
-    }
-
-    /**
-     * Data provider for incorrect translation class
-     *
-     * @return array
-     */
-    public static function incorrectClassNameDataProvider()
-    {
-        return [
-            'incorrect_translation_class' => [
-                [
-                    'translation_class' => 'translation_class',
-                    'entity_class'      => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\Item',
-                    'field'             => 'name'
-                ]
-            ]
-        ];
-    }
-
-    /**
-     * Test with incorrect translation and/or entity classes
-     *
-     * @param array $options
-     *
-     * @dataProvider incorrectClassNameDataProvider
-     * @expectedException \ReflectionException
-     * @expectedExceptionMessage Class translation_class does not exist
-     */
-    public function testReflectionException(array $options)
-    {
-        $target             = $this->getTargetedClass($options);
-        $form               = $this->getFormMock();
-        $translatableEntity = $this->getTranslatableEntityMock();
-        $event              = $this->getEventMock($form, $translatableEntity, []);
-
-        $target->preSetData($event);
-    }
-
-    /**
-     * Data provider for getSubsbriberEvents method
-     *
-     * @return array
-     */
-    public static function getSubscriberEventsDataProvider()
-    {
-        return [
-            [
-                [
-                    'entity_class'      => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\Item',
-                    'translation_class' => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\ItemTranslation',
-                    'field'             => 'name'
-                ]
-            ]
-        ];
-    }
-
-    /**
-     * Test subscriber events
-     *
-     * @param array $options
-     *
-     * @dataProvider getSubscriberEventsDataProvider
-     */
-    public function testGetSubscriberEvents(array $options)
-    {
-        $target = $this->getTargetedClass($options);
-        $events = $target->getSubscribedEvents();
-
-        $this->assertTrue(array_key_exists('form.pre_set_data', $events), 'preSetData');
-        $this->assertTrue(array_key_exists('form.post_bind', $events), 'postBind');
-        $this->assertTrue(array_key_exists('form.bind', $events), 'bind');
-    }
-
-    /**
-     * Test preSet without data
-     * It should do nothing
-     *
-     * @param array $options
-     *
-     * @dataProvider getSubscriberEventsDataProvider
-     */
-    public function testPreSetWithoutData(array $options)
-    {
-        $target = $this->getTargetedClass($options);
-        $form   = $this->getFormMock();
-        $event  = $this->getEventMock($form);
-
-        $form->expects($this->never())
-             ->method('getParent');
-
-        $target->preSetData($event);
-    }
-
-    /**
-     * Data provider for preSet data
-     *
-     * @return array
-     */
-    public static function preSetDataProvider()
-    {
-        return [
-            'classic' => [
-                [
-                    'entity_class'      => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\Item',
-                    'translation_class' => 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\ItemTranslation',
-                    'field'             => 'name',
-                    'required_locale'   => [],
-                    'locales'           => ['en_US', 'fr_FR'],
-                    'widget'            => 'text'
-                ]
-            ]
-        ];
-    }
-
-    /**
-     * Test preSetData
-     * It should add form fields for each translations
-     *
-     * @param array $options
-     *
-     * @dataProvider preSetDataProvider
-     */
-    public function testPreSetData(array $options)
-    {
-        $target                = $this->getTargetedClass($options);
-        $translatableEntity    = $this->getTranslatableEntityMock();
-
-        $event = $this->getEventMock($this->form, $translatableEntity, []);
-
-        $locales = $options['locales'];
-        $requiredLocales = $options['required_locale'];
-
-        foreach ($locales as $index => $locale) {
-            $this->formFactory->expects($this->at($index))
-                ->method('createNamed')
-                ->with(
-                    $locale,
-                    $options['widget'],
-                    '',
-                    [
-                        'label'           => $this->localeConfig['locales'][$locale]['label'],
-                        'required'        => in_array($locale, $requiredLocales),
-                        'mapped'          => false,
-                        'auto_initialize' => false
-                    ]
-                )
-                ->will($this->returnValue($field = $this->getFormMock()));
-
-            $this->form->expects($this->any())
-                 ->method('add')
-                 ->with($this->equalTo($field));
-        }
-
-        $target->preSetData($event);
-    }
-
-    /**
-     * Data provider for data binding
-     *
-     * @return array
-     */
-    public static function bindDataProvider()
-    {
-        return self::preSetDataProvider();
-    }
-
-    /**
-     * Test data binding
-     * It should validate required translations
-     *
-     * @param array $options
-     *
-     * @dataProvider bindDataProvider
-     */
-    public function testBindData(array $options)
-    {
-        $target                = $this->getTargetedClass($options);
-        $translatableEntity    = $this->getTranslatableEntityMock();
-
-        $locales = $options['locales'];
-
-        foreach ($locales as $index => $locale) {
-            $this->form->expects($this->at($index))
-                 ->method('get')
-                 ->with($locale)
-                 ->will($this->returnValue($defaultField = $this->getFormMock()));
-
-            $defaultField->expects($this->any())
-                         ->method('getData')
-                         ->will($this->returnValue(null));
-        }
-
-        $event = $this->getEventMock($this->form, $translatableEntity, []);
-
-        $target->submit($event);
-    }
-
-    /**
-     * Data provider post data binding
-     * It should add translation after binding
-     *
-     * @return array
-     */
-    public static function postBindDataProvider()
-    {
-        return self::preSetDataProvider();
-    }
-
-    /**
-     * Test data post binding
-     * It should add translation if content is provided
-     *
-     * @param array $options
-     *
-     * @dataProvider postBindDataProvider
-     */
-    public function testPostBindData(array $options)
-    {
-        $this->markTestIncomplete('Not totally implemented');
-
-        $target = $this->getTargetedClass($options);
-        $translatableEntity = $this->getTranslatableEntityMock();
-
-        $event = $this->getEventMock(
-            $this->form,
-            $translatableEntity
-        );
-
-        $target->postBind($event);
-    }
-
-    /**
-     * Create tested subscriber
-     *
-     * @param array $options
-     *
-     * @return \Pim\Bundle\TranslationBundle\Form\Subscriber\AddTranslatableFieldSubscriber
-     */
-    protected function getTargetedClass(array $options)
-    {
-        $options['locale_config'] = $this->localeConfig;
-
-        return new AddTranslatableFieldSubscriber(
-            $this->formFactory,
-            $this->getValidatorMock(),
-            $this->getUserContextMock(),
-            $this->getLocaleHelperMock(),
-            $options
-        );
-    }
-
-    /**
-     * Get LocaleHelperMock
-     *
-     * @return \Pim\Bundle\CatalogBundle\Helper\LocaleHelper
-     */
-    protected function getLocaleHelperMock()
-    {
-        $helper = $this->getMockBuilder('Pim\Bundle\CatalogBundle\Helper\LocaleHelper')
-            ->disableOriginalConstructor()
-            ->getMock();
-        $helper->expects($this->any())
-            ->method('getLocaleLabel')
-            ->will($this->returnArgument(0));
-
-        return $helper;
-    }
-
-    /**
-     * Get UserContext mock
-     *
-     * @return \Pim\Bundle\UserBundle\Context\UserContext
-     */
-    protected function getUserContextMock()
-    {
-        $userContext = $this->getMockBuilder('Pim\Bundle\UserBundle\Context\UserContext')
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $userContext->expects($this->any())
-            ->method('getUserLocaleCodes')
-            ->will($this->returnValue(['en_US', 'fr_FR']));
-
-        return $userContext;
-    }
-
-    /**
-     * Create FormFactory mock
-     *
-     * @return Mock
-     */
-    protected function getFormFactoryMock()
-    {
-        $childForm = $this->getFormMock();
-
-        $formFactory = $this->getMockBuilder('Symfony\Component\Form\FormFactory')
-                            ->disableOriginalConstructor()
-                            ->setMethods(['createNamed'])
-                            ->getMock();
-
-        $formFactory->expects($this->any())
-                    ->method('createNamed')
-                    ->will($this->returnValue($childForm));
-
-        return $formFactory;
-    }
-
-    /**
-     * Create Validator mock
-     *
-     * @return Mock
-     */
-    protected function getValidatorMock()
-    {
-        return $this->getMock('Symfony\Component\Validator\Validator\ValidatorInterface');
-    }
-
-    /**
-     * Create FormEvent mock
-     *
-     * @param Form  $form
-     * @param array $parentData
-     * @param array $data
-     *
-     * @return Mock
-     */
-    protected function getEventMock($form, $parentData = null, array $data = null)
-    {
-        $event = $this
-            ->getMockBuilder('Symfony\Component\Form\FormEvent')
-            ->disableOriginalConstructor()
-            ->setMethods(['getData', 'getForm'])
-            ->getMock();
-
-        $event->expects($this->any())
-              ->method('getData')
-              ->will($this->returnValue($data));
-
-        $event->expects($this->any())
-              ->method('getForm')
-              ->will($this->returnValue($form));
-
-        $parentForm = $this->getFormMock();
-
-        $form->expects($this->any())
-             ->method('getParent')
-             ->will($this->returnValue($parentForm));
-
-        $parentForm->expects($this->any())
-                   ->method('getData')
-                   ->will($this->returnValue($parentData));
-
-        return $event;
-    }
-
-    /**
-     * Create Form mock
-     *
-     * @return Mock
-     */
-    protected function getFormMock()
-    {
-        $form = $this->getMockBuilder('Symfony\Component\Form\Form')
-                     ->disableOriginalConstructor()
-                    ->setMethods(['getParent', 'getData', 'add', 'get', 'addError'])
-                    ->getMock();
-
-        $form->expects($this->any())
-             ->method('add')
-             ->will($this->returnValue('Symfony\Component\Form\FormInterface'));
-
-        if ($this->form) {
-            $form->expects($this->any())
-                 ->method('get')
-                 ->will($this->returnValue($this->form));
-        }
-
-        return $form;
-    }
-
-    /**
-     * Get translatable entity mock
-     *
-     * @return Mock
-     */
-    protected function getTranslatableEntityMock()
-    {
-        return $this->getMockForAbstractClass('Pim\Bundle\TranslationBundle\Tests\Entity\Item');
-    }
-
-    /**
-     * Get translated entity mock
-     *
-     * @return Mock
-     */
-    protected function getTranslationMock()
-    {
-        return $this->getMockForAbstractClass('Pim\Bundle\TranslationBundle\Tests\Entity\ItemTranslation');
-    }
-}
