--- src/Pim/Bundle/TranslationBundle/Tests/Unit/Form/Type/TranslatableFieldTypeTest.php
@@ -1,209 +1,0 @@
-<?php
-
-namespace Pim\Bundle\TranslationBundle\Tests\Unit\Form\Type;
-
-use Pim\Bundle\TranslationBundle\Form\Type\TranslatableFieldType;
-use Symfony\Component\Form\Extension\Validator\Type\FormTypeValidatorExtension;
-use Symfony\Component\Form\Forms;
-use Symfony\Component\Form\Test\TypeTestCase;
-
-/**
- * Test related class
- *
- * @author    Romain Monceau <romain@akeneo.com>
- * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-class TranslatableFieldTypeTest extends TypeTestCase
-{
-    /**
-     * @var TranslatableFieldType
-     */
-    protected $type;
-
-    /**
-     * @var \Symfony\Component\Form\FormInterface
-     */
-    protected $form;
-
-    /**
-     * @var array
-     */
-    protected $options;
-
-    /**
-     * @staticvar string
-     */
-    const OPT_NAME = 'name';
-
-    /**
-     * @staticvar string
-     */
-    const OPT_ENTITY_CLASS = 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\Item';
-
-    /**
-     * @staticvar string
-     */
-    const OPT_TRANSLATION_CLASS = 'Pim\\Bundle\\TranslationBundle\\Tests\\Entity\\ItemTranslation';
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function setUp()
-    {
-        parent::setUp();
-
-        // redefine form factory and builder to add translatable field
-        $this->builder->add('pim_translatable_field');
-        $this->factory = Forms::createFormFactoryBuilder()
-            ->addTypeExtension(
-                new FormTypeValidatorExtension(
-                    $this->getMock('Symfony\Component\Validator\Validator\ValidatorInterface')
-                )
-            )
-            ->addType(
-                new TranslatableFieldType(
-                    $this->getMock('Symfony\Component\Validator\Validator\ValidatorInterface'),
-                    $this->getUserContextMock(),
-                    $this->getLocaleHelperMock()
-                )
-            )
-            ->getFormFactory();
-
-        // Create form type
-        $this->type = new TranslatableFieldType(
-            $this->getMock('Symfony\Component\Validator\Validator\ValidatorInterface'),
-            $this->getUserContextMock(),
-            $this->getLocaleHelperMock()
-        );
-        $this->options = $this->buildOptions(self::OPT_ENTITY_CLASS, self::OPT_NAME, self::OPT_TRANSLATION_CLASS);
-
-        $this->form = $this->factory->create($this->type, null, $this->options);
-    }
-
-    /**
-     * Create mock for user context
-     *
-     * @return \Pim\Bundle\UserBundle\Context\UserContext
-     */
-    protected function getUserContextMock()
-    {
-        $userContext = $this
-            ->getMockBuilder('Pim\Bundle\UserBundle\Context\UserContext')
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $userContext->expects($this->any())
-            ->method('getUserLocaleCodes')
-            ->will($this->returnValue(['en_US', 'fr_FR']));
-
-        return $userContext;
-    }
-
-    /**
-     * Test method related
-     */
-    public function testFormCreate()
-    {
-        // Assert default options
-        $expectedOptions = [
-            'entity_class'      => false,
-            'field'             => false,
-            'locales'           => ['en_US', 'fr_FR'],
-            'required_locale'   => [],
-            'translation_class' => false,
-            'widget'            => 'text'
-        ];
-
-        // Assert options
-        $options = $this->form->getConfig()->getOptions();
-        $expectedOptions['entity_class']      = self::OPT_ENTITY_CLASS;
-        $expectedOptions['field']             = self::OPT_NAME;
-        $expectedOptions['translation_class'] = self::OPT_TRANSLATION_CLASS;
-        foreach ($expectedOptions as $expectedKey => $expectedValue) {
-            $this->assertArrayHasKey($expectedKey, $options);
-            $this->assertEquals($expectedValue, $options[$expectedKey]);
-        }
-
-        // Assert name
-        $this->assertEquals('pim_translatable_field', $this->form->getName());
-    }
-
-    /**
-     * Create options
-     *
-     * @param string $entityClass      Entity class name
-     * @param string $fieldName        Entity field name
-     * @param string $translationClass Translation class name
-     *
-     * @return string[]
-     */
-    protected function buildOptions($entityClass, $fieldName, $translationClass)
-    {
-        return [
-            'entity_class'      => $entityClass,
-            'field'             => $fieldName,
-            'translation_class' => $translationClass
-        ];
-    }
-
-    /**
-     * Data provider for options
-     *
-     * @return array
-     */
-    public static function dataOptionsProvider()
-    {
-        return [
-            [self::OPT_ENTITY_CLASS, self::OPT_NAME, null],
-            [self::OPT_ENTITY_CLASS, null, self::OPT_TRANSLATION_CLASS],
-            [null, self::OPT_NAME, self::OPT_TRANSLATION_CLASS]
-        ];
-    }
-
-    /**
-     * Assert buildForm exceptions
-     *
-     * @param string $entityClass      Entity class name
-     * @param string $fieldName        Entity field name
-     * @param string $translationClass Translation class name
-     *
-     * @dataProvider dataOptionsProvider
-     * @expectedException \Symfony\Component\Form\Exception\InvalidConfigurationException
-     */
-    public function testAssertException($entityClass, $fieldName, $translationClass)
-    {
-        $options = $this->buildOptions($entityClass, $fieldName, $translationClass);
-        $this->factory->create($this->type, null, $options);
-    }
-
-    /**
-     * Get LocaleHelperMock
-     *
-     * @return \Pim\Bundle\CatalogBundle\Helper\LocaleHelper
-     */
-    protected function getLocaleHelperMock()
-    {
-        $helper = $this->getMockBuilder('Pim\Bundle\CatalogBundle\Helper\LocaleHelper')
-            ->disableOriginalConstructor()
-            ->getMock();
-        $helper->expects($this->any())
-            ->method('getLocaleLabel')
-            ->will($this->returnArgument(0));
-
-        return $helper;
-    }
-
-    /**
-     * Get ACL SecurityFacade mock
-     *
-     * @return \Oro\Bundle\SecurityBundle\SecurityFacade
-     */
-    protected function getSecurityFacadeMock()
-    {
-        return $this
-            ->getMockBuilder('Oro\Bundle\SecurityBundle\SecurityFacade')
-            ->disableOriginalConstructor()
-            ->getMock();
-    }
-}
